Add Basic repo structure based on Sphinx
This is the initial proposal for the repo structure using Sphinx as a
tool to generate documentation.

The documentation that we will be writing will be in the docs/source
directory.

A requirements.txt was added to make the replication of the environment
needed to build this easier.

NOTE: the following files were automatically generated by the Sphnix
tool:

* docs/Makefile
* docs/source/conf.py
* docs/soruce.index.rst

Change-Id: I6ccd57a6e1a274bb396999ce272678fb9da6f702
Signed-off-by: Juan Antonio Osorio Robles <juan.osorio.robles@ericsson.com>